Divide main logic with add Inventory class
